1. Refactoring packages
