--- src/main/java/org/bahmni/eventrouterservice/configuration/ServiceName.java
@@ -1,5 +1,0 @@
-package org.bahmni.eventrouterservice.configuration;
-
-public enum ServiceName {
-    BAHMNI, GCP
-}
