+++ src/main/java/org/bahmni/eventrouterservice/model/ServiceName.java
@@ -1,0 +1,5 @@
+package org.bahmni.eventrouterservice.model;
+
+public enum ServiceName {
+    BAHMNI, GCP
+}
