+++ src/main/java/org/bahmni/eventrouterservice/publisher/bahmni/BahmniEventPublisher.java
@@ -1,6 +1,6 @@
 package org.bahmni.eventrouterservice.publisher.bahmni;
 
-import org.bahmni.eventrouterservice.publisher.common.exception.FailedToPublishException;
+import org.bahmni.eventrouterservice.publisher.exception.FailedToPublishException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
