+++ src/main/java/org/bahmni/eventrouterservice/publisher/configuration/PublisherConfiguration.java
@@ -3,11 +3,11 @@ package org.bahmni.eventrouterservice.publisher.configuration;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.bahmni.eventrouterservice.configuration.ServiceName;
-import org.bahmni.eventrouterservice.configuration.Topic;
+import org.bahmni.eventrouterservice.model.ServiceName;
+import org.bahmni.eventrouterservice.model.Topic;
 import org.bahmni.eventrouterservice.exception.FailedToLoadConfiguration;
-import org.bahmni.eventrouterservice.publisher.common.exception.EndpointNotConfiguredForPublisherException;
-import org.bahmni.eventrouterservice.publisher.common.exception.TopicNotConfiguredForPublisherException;
+import org.bahmni.eventrouterservice.publisher.exception.EndpointNotConfiguredForPublisherException;
+import org.bahmni.eventrouterservice.publisher.exception.TopicNotConfiguredForPublisherException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
