+++ src/main/java/org/bahmni/eventrouterservice/publisher/exception/PublisherNotConfiguredException.java
@@ -1,4 +1,4 @@
-package org.bahmni.eventrouterservice.publisher.common.exception;
+package org.bahmni.eventrouterservice.publisher.exception;
 
 public class PublisherNotConfiguredException extends RuntimeException {
     public PublisherNotConfiguredException(String publisherId) {
