+++ src/main/java/org/bahmni/eventrouterservice/publisher/exception/TopicNotConfiguredForPublisherException.java
@@ -1,4 +1,4 @@
-package org.bahmni.eventrouterservice.publisher.common.exception;
+package org.bahmni.eventrouterservice.publisher.exception;
 
 public class TopicNotConfiguredForPublisherException extends RuntimeException {
     public TopicNotConfiguredForPublisherException(String publisherId) {
