+++ src/main/java/org/bahmni/eventrouterservice/publisher/gcp/GCPEventPublisher.java
@@ -1,7 +1,7 @@
 package org.bahmni.eventrouterservice.publisher.gcp;
 
 import com.google.cloud.spring.pubsub.core.PubSubTemplate;
-import org.bahmni.eventrouterservice.publisher.common.exception.FailedToPublishException;
+import org.bahmni.eventrouterservice.publisher.exception.FailedToPublishException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +9,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.CancellationException;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
 @ConditionalOnProperty(
