+++ src/main/java/org/bahmni/eventrouterservice/publisher/service/EventPublisherService.java
@@ -1,4 +1,4 @@
-package org.bahmni.eventrouterservice.publisher.common.service;
+package org.bahmni.eventrouterservice.publisher.service;
 
 public interface EventPublisherService {
     void publish(String payload, String publisherId);
