+++ src/main/java/org/bahmni/eventrouterservice/subscriber/exception/FailedToSubscribeException.java
@@ -1,4 +1,4 @@
-package org.bahmni.eventrouterservice.subscriber.common.exception;
+package org.bahmni.eventrouterservice.subscriber.exception;
 
 public class FailedToSubscribeException extends RuntimeException {
     public FailedToSubscribeException(Exception exception) {
