+++ src/test/java/org/bahmni/eventrouterservice/publisher/common/exception/PublisherNotConfiguredExceptionTest.java
@@ -1,5 +1,6 @@
 package org.bahmni.eventrouterservice.publisher.common.exception;
 
+import org.bahmni.eventrouterservice.publisher.exception.PublisherNotConfiguredException;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
