+++ src/test/java/org/bahmni/eventrouterservice/publisher/common/service/EventPublisherServiceFactoryTest.java
@@ -3,10 +3,12 @@ package org.bahmni.eventrouterservice.publisher.common.service;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.bahmni.eventrouterservice.publisher.bahmni.BahmniEventPublisherService;
-import org.bahmni.eventrouterservice.publisher.common.exception.PublisherNotConfiguredException;
+import org.bahmni.eventrouterservice.publisher.exception.PublisherNotConfiguredException;
 import org.bahmni.eventrouterservice.publisher.configuration.PublisherConfiguration;
 import org.bahmni.eventrouterservice.publisher.configuration.PublisherConfiguration.PublisherDescription;
 import org.bahmni.eventrouterservice.publisher.gcp.GCPEventPublisherService;
+import org.bahmni.eventrouterservice.publisher.service.EventPublisherService;
+import org.bahmni.eventrouterservice.publisher.service.EventPublisherServiceFactory;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
