+++ src/test/java/org/bahmni/eventrouterservice/publisher/gcp/GCPEventPublisherTest.java
@@ -1,7 +1,7 @@
 package org.bahmni.eventrouterservice.publisher.gcp;
 
 import com.google.cloud.spring.pubsub.core.PubSubTemplate;
-import org.bahmni.eventrouterservice.publisher.common.exception.FailedToPublishException;
+import org.bahmni.eventrouterservice.publisher.exception.FailedToPublishException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
