+++ src/test/java/org/bahmni/eventrouterservice/subscriber/gcp/GCPEventSubscriberTest.java
@@ -7,8 +7,8 @@ import com.google.cloud.spring.pubsub.support.AcknowledgeablePubsubMessage;
 import com.google.protobuf.ByteString;
 import com.google.pubsub.v1.PubsubMessage;
 import org.bahmni.eventrouterservice.publisher.bahmni.BahmniEventPublisherService;
-import org.bahmni.eventrouterservice.publisher.common.service.EventPublisherServiceFactory;
-import org.bahmni.eventrouterservice.subscriber.common.exception.FailedToSubscribeException;
+import org.bahmni.eventrouterservice.publisher.service.EventPublisherServiceFactory;
+import org.bahmni.eventrouterservice.subscriber.exception.FailedToSubscribeException;
 import org.bahmni.eventrouterservice.subscriber.configuration.SubscriberConfiguration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -22,7 +22,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
-import static org.bahmni.eventrouterservice.configuration.ServiceName.GCP;
+import static org.bahmni.eventrouterservice.model.ServiceName.GCP;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
